feat: refactored dependencies from bi-directional to uni-directional
