+++ imports/api/rest/put-pending-invitations.js
@@ -3,9 +3,8 @@ import { invite } from '../../util/email-invite'
 import { Meteor } from 'meteor/meteor'
 
 import { reloadCaseFields } from '../cases'
-import UnitRolesData from '../unit-roles-data'
+import UnitRolesData, { defaultRoleVisibility } from '../unit-roles-data'
 import { logger } from '../../util/logger'
-import { defaultRoleVisibility } from '../units'
 
 export default (req, res) => {
   if (req.query.accessToken !== process.env.API_ACCESS_TOKEN) {
