+++ imports/api/unit-roles-data.js
@@ -6,10 +6,14 @@ import { check } from 'meteor/check'
 import { HTTP } from 'meteor/http'
 import { Random } from 'meteor/random'
 import randToken from 'rand-token'
-import { addUserToRole, defaultRoleVisibility } from './units'
 import { findOrCreateUser } from './custom-users'
 import UnitMetaData from './unit-meta-data'
-import PendingInvitations, { KEEP_DEFAULT, REMOVE_USER, REPLACE_DEFAULT, collectionName as pendingInvitaitonsCollName } from './pending-invitations'
+import PendingInvitations, {
+  KEEP_DEFAULT,
+  REMOVE_USER,
+  REPLACE_DEFAULT,
+  collectionName as pendingInvitationsCollName
+} from './pending-invitations'
 import unitUserInvitedTemplate from '../email-templates/unit-user-invited'
 import { logger } from '../util/logger'
 import { getIncrementFor } from './increment-counters'
@@ -40,6 +44,23 @@ export const possibleRoles = [
   {
     name: roleEnum.AGENT
   }
+]
+
+export const defaultRoleVisibility = {
+  [roleEnum.TENANT]: true,
+  [roleEnum.OWNER_LANDLORD]: true,
+  [roleEnum.CONTRACTOR]: true,
+  [roleEnum.MGT_COMPANY]: true,
+  [roleEnum.AGENT]: true,
+  'Occupant': true
+}
+
+export const roleSortOrder = [
+  roleEnum.TENANT,
+  roleEnum.OWNER_LANDLORD,
+  roleEnum.AGENT,
+  roleEnum.MGT_COMPANY,
+  roleEnum.CONTRACTOR
 ]
 
 const UnitRolesData = new Mongo.Collection(collectionName)
@@ -77,6 +98,115 @@ export function inviteUserToRole (invitorId, unitMongoId, inviteeUser, roleType,
   return { accessToken }
 }
 
+export const addUserToRole = (
+  invitingUser, inviteeUser, unitBzId, role, invType, isOccupant, errorLogParams = {}, doLiveUpdate, isVisible = true,
+  isDefaultInvited = true, roleVisibility = defaultRoleVisibility
+) => {
+  // Filling up role visibility in case some of it is missing
+  roleVisibility = Object.assign({}, defaultRoleVisibility, roleVisibility)
+
+  // Creating matching invitation records
+  const invitationObj = {
+    invitedBy: invitingUser.bugzillaCreds.id,
+    invitee: inviteeUser.bugzillaCreds.id,
+    mefeInvitationIdIntValue: getIncrementFor(pendingInvitationsCollName),
+    type: invType,
+    unitId: unitBzId,
+    role,
+    isOccupant
+  }
+
+  // TODO: Once all dependencies of role resolving are moved from invitations to UnitRolesData, remove this
+  // Creating the invitation as pending first
+  const invitationId = PendingInvitations.insert(invitationObj)
+
+  // Linking invitation to user
+  Meteor.users.update(inviteeUser._id, {
+    $push: {
+      receivedInvites: {
+        unitId: invitationObj.unitId,
+        invitedBy: invitingUser._id,
+        timestamp: Date.now(),
+        type: invitationObj.type,
+        invitationId,
+        role,
+        isOccupant
+      }
+    }
+  })
+
+  // Adding to the user to a role on BZ using lambda
+  try {
+    HTTP.call('POST', process.env.INVITE_LAMBDA_URL, {
+      data: [Object.assign({ _id: invitationId }, invitationObj)],
+      headers: {
+        Authorization: `Bearer ${process.env.API_ACCESS_TOKEN}`
+      }
+    })
+  } catch (e) {
+    logger.error({
+      ...errorLogParams,
+      step: 'INVITE lambda request, unit cleanup might be necessary',
+      error: e
+    })
+    throw new Meteor.Error('Invite API Lambda error', e, { lambdaStatusCode: e.response.statusCode })
+  }
+
+  // Marking the pending invitation as "done", now that the API responded with success
+  PendingInvitations.update({ _id: invitationId }, {
+    $set: {
+      done: true
+    }
+  })
+  Meteor.users.update({
+    _id: inviteeUser._id,
+    'receivedInvites.invitationId': invitationId
+  }, {
+    $set: {
+      'receivedInvites.$.done': true
+    }
+  })
+
+  // Updating the roles collection to sync with BZ's state
+  const unitRoleQuery = {
+    roleType: role,
+    unitBzId
+  }
+
+  const unitRoleModifier = {
+    $push: {
+      members: {
+        id: inviteeUser._id,
+        isVisible,
+        isDefaultInvited,
+        isOccupant,
+        roleVisibility
+      }
+    }
+  }
+
+  UnitRolesData.update(unitRoleQuery, unitRoleModifier)
+
+  // Matching the role if the defaultAssigneeId is not defined and sets it to the current user. Does nothing otherwise
+  let doForceAssigneeUpdate
+  switch (invType) {
+    case REPLACE_DEFAULT:
+      doForceAssigneeUpdate = true
+      break
+    default:
+      doForceAssigneeUpdate = false
+  }
+  const assigneeUpdateQuery = doForceAssigneeUpdate ? unitRoleQuery : {
+    defaultAssigneeId: -1,
+    ...unitRoleQuery
+  }
+  UnitRolesData.update(assigneeUpdateQuery, {
+    $set: {
+      defaultAssigneeId: inviteeUser._id
+    }
+  })
+}
+
 export function removeRoleMember (requestorId, unitBzId, email, errorLogParams) {
   const unitMeta = UnitMetaData.findOne({ bzId: unitBzId })
   const unitRoles = UnitRolesData.find({ unitBzId }).fetch()
@@ -102,7 +232,7 @@ export function removeRoleMember (requestorId, unitBzId, email, errorLogParams) 
   const invitationObj = {
     invitedBy: requestorUser.bugzillaCreds.id,
     invitee: userToRemove.bugzillaCreds.id,
-    mefeInvitationIdIntValue: getIncrementFor(pendingInvitaitonsCollName),
+    mefeInvitationIdIntValue: getIncrementFor(pendingInvitationsCollName),
     type: REMOVE_USER,
     unitId: unitBzId,
     role: toRemoveRole.roleType,
